requirement.txt
pytorch_lightning
torchvision